requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pypdf>=5.0
python-docx>=1.1
redis>=5.0
python-jose[cryptography]>=3.3
pydantic>=2.7
pydantic-settings>=2.2
python-dotenv>=1.0
openai>=1.0
numpy>=1.26
langgraph>=0.2
langchain-core>=0.3
langchain-openai>=0.2

# QA / Dev tooling
pytest>=8.0
pytest-cov>=5.0
pytest-watch>=4.2
ruff>=0.6
black>=24.0
mypy>=1.10
httpx>=0.27

